Hook change to component
