+++ src/components/ui/media/Image.tsx
@@ -9,7 +9,7 @@ import {
   StyleSheet,
   useWindowDimensions,
 } from 'react-native'
-import {useNavigateToImageViewer} from '@/hooks/useNavigateToImageViewer'
+import {NavigateToImageViewer} from '@/components/ui/media/NavigateToImageViewer'
 import {Theme} from '@/themes/themes'
 import {ImageAspectRatio} from '@/themes/tokens/media'
 import {useThemable} from '@/themes/useThemable'
@@ -44,11 +44,6 @@ export const Image = ({
 }: ImageProps) => {
   const {height: windowHeight, width: windowWidth} = useWindowDimensions()
   const [width, setWidth] = useState<number | undefined>(undefined)
-  const Wrapper = useNavigateToImageViewer(openInImageViewer, {
-    ...imageProps,
-    source,
-    aspectRatio,
-  })
 
   const createdStyles = useMemo(
     () => createStyles(aspectRatio, width),
@@ -70,7 +65,9 @@ export const Image = ({
   )
 
   return (
-    <Wrapper>
+    <NavigateToImageViewer
+      imageProps={{...imageProps, source, aspectRatio}}
+      shouldNavigate={openInImageViewer}>
       <ImageRN
         accessibilityIgnoresInvertColors
         accessibilityLanguage="nl-NL"
@@ -79,7 +76,7 @@ export const Image = ({
         style={[styles.image, style]}
         {...imageProps}
       />
-    </Wrapper>
+    </NavigateToImageViewer>
   )
 }
